+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 -   This project adheres to [Semantic Versioning][semver].
 -   This project uses [Gitmoji][gitmoji]
 
+## 3.1.0
+
+-   🏗️ some changes in `count`, `every`, `filter`, `some`,
+    `handleCurriedIsomorphicIterable`, `handleIsomorphicIterable` and
+    `IsomorphicGeneratorFunction` types.
+-   ⬆️ dependency upgrade.
+
 ## 3.0.12
 
 -   ⬆️ dependency upgrade.